requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0

# Media Processing
ffmpeg-python==0.2.0

# AI & ML
openai-whisper==20231117
google-generativeai==0.3.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# Utilities
numpy==1.24.3
torch>=2.1.0